Implement softmax function and enhance TimeStep and BehaviourTracer classes
A softmax function is added to handle zero probabilities in the action distribution. The TimeStep class is expanded to include several new attributes related to action and control information. BehaviourTracer class is also enhanced; it now has methods to add new time steps to the trajectory and to summarize all time steps upon completion of the trajectory.