Update random agent start position handling
The agent's start position and direction initialization has been refactored. The code now allows agent_start_pos and agent_start_dir to be None, representing a random initialization. The reset method has been removed. As part of this change, a new function find_empty_position was introduced which is used to find a random start position if needed. The width and height of the environment are now also defined based on the environment layout.