Move data to CPU and add random data rotation
Updated the replay_buffer to store data on the CPU instead of GPU for better memory management. Also introduced a functionality to randomly rotate the state data. This random rotation adds more variety to the training process, potentially enhancing the robustness of the model by providing it with a wider range of scenarios.