Add method to manage prioritized replay buffer in PPO
A new method was added to handle discrete prioritized replay buffer charging in the PPO algorithm file. In addition to this, we have refactored the method of tensor repetition, moving it from being a local to a global function. These changes contribute to more efficient memory management and clearer code structure.