Refactor DQN implementation to use FlexibleImageEncoder
Refactored the implementation of DQN to use 'FlexibleImageEncoder' exclusively for action selection and prediction. The class 'QModel' and associated code for its use have been commented out. Uses of the 'encoder' object is also removed throughout the code, shifting its functionality to the 'q_model' object and eliminating the need for an additional encoder optimizer.