Normalize observation tensor in simple_gridworld
Changes in simple_gridworld.py are dedicated to the normalization of the observation tensor. Observation tensor is now converted to float32 type, and its values are rescaled to range from 0 to 1. This normalization could help providing more standardized inputs to further models or procedures, leading to potential improvements in performance and stability.