Add random grid rotation and flip in TextGridWorld
In class TextGridWorld, a new boolean parameter is added named 'make_random'. If enabled, it generates random grid rotations and flipping. Along with this, an update has been made to the file path in CustomEnvFromFile call to include the 'envs/simple_grid' directory. This enhances the randomness and complexity of the grid-world, potentially helping to train more robust agents.