Introduce max_steps parameter to SimpleGridworld
The SimpleGridworld environment now takes an additional 'max_steps' parameter that sets an upper limit to the number of steps that can be taken in one episode. This new parameter is checked in the `step` function, and terminates the episode if the maximum number of steps is reached. All environment configurations in `train_baseline.py` have been updated accordingly and now include a default value for 'max_steps'.