Adjust make_env return type and enhance find_newest_model function
The function make_env now returns a SimpleGridWorld instance instead of gymnasium.Env. In the function find_newest_model, the code is improved to return not only the latest model's path but also its iteration number. This is useful in the main loop where the training starts from the iteration number of the previous stored model.