Add plot_actions method for detailed action visualization in gridworlds
This update introduces the `plot_actions` method to provide more detailed visualizations of agent's and prior actions in gridworld environments. The method uses colored arrows to indicate direction of actions and their probabilities. Minor refactoring was also done to the calculation of the delta control info tensor and the `plot_grid` method.