Handle None value in latest_model_iteration in train_baseline.py
In train_baseline.py, a condition has been introduced to check if the latest_model_iteration is None. To avoid breaking the execution of the program, if this condition is met, the latest_model_iteration will be assigned a default value of -1. This occurs when there are no iterations for the latest model