Add option to save directed graph in simple_gridworld
This commit introduces an option for the user to save the directed graph generated by simple_gridworld to an image file. This allows the user to view the graph layout at their convenience without re-running the script. It also simplifies further analysis of the graph by providing a persisting visual documentation.