Add 'show' parameter to 'make_directed_graph' function in simple_gridworld.py
This update modifies the 'make_directed_graph' function to include a 'show' parameter that controls the display of the directed graph in the system's window. This parameter is set to false by default but can be set to true when we want to display the graph. Additionally, the environment for creating the graph is changed to 'gridworld-empty-7.txt'.