Import PPO and modify model updating mechanism
Added import statement for PPO from stable_baselines3 in utils.py. Modified the model updating mechanism to instantiate a new PPO object with parameters copied from the current model. Removed the deeper copying of the model and usage of session_name in the logging process in favor of a simpler and more efficient approach.