Add randomness and adjustment frequencies to training environment
The update incorporates random selection to determine whether the original environment or an abstract version is returned. Additionally, update frequencies for the environment and agent are adjusted, and the abstract rate has been introduced to the environment's callback function. These changes aim to optimize and bring variability in the training process.