Add environment reset option in behaviour_tracer
The update adds an option to reset the environment in the behaviour_tracer constructor. An additional gridworld configuration has also been added to the list of environments, with unique properties such as predetermined goal_position and make_random set to False. These changes offer more flexibility and control over the environment during the training and testing process.