Add function to get traversed grids and avoid traps
A function `get_traversed_grids` was added to determine the grids traversed from a previous position to a current position in the gridworld. Moreover, checks were introduced to avoid next actions that would cause the agent to fly over a trap or a wall. This updates the behavior of the agent to be more cautious while still aspiring to take the optimal path. Changes in the clustering parameters were also included.