Adjust training parameters and uncomment simple gridworld maze environments
The training parameters in 'train_baseline_abs.py' have been adjusted; the number of clusters ending range was reduced to 40, the abstraction rate is now 0.5 and total timesteps set to 300000. Additionally, previously commented-out environments for maze training and testing in 'configurations.py' have been uncommented, enabling a new maze environment for the agent to learn and test in.