Implement image reconstruction in feature model training
In this update, an image reconstruction functionality has been integrated into the feature model training process. This was achieved by introducing a decoder part in the model, FlexibleImageDecoder, and updating the training loss to include pixel loss. In addition, we have also added assertion to ensure that the observation size configuration is consistent. Lastly, reconstructed images are plotted and saved at each testing frequency.