Reduce num_hidden_channels and adjust layers in FlexibleImageDecoder
The number of hidden channels in the FlexibleImageDecoder class within feature_model.py has been reduced from 128 to 32. This change modifies the initialisation, along with updates to the fully connected layers, convolution and batch normalization layers. These alterations are expected to reduce the complexity of the feature learning, potentially improving speed and efficiency.