Uncomment `env_configurations` variable and refactor plotting method
The `env_configurations` variable has been uncommented and added back in the UpdateFeatureExtractorCallback class. The commented code related to the prior agent and the updating of the environment has been removed. The plotting method has been refactored to get the environment name from the configuration file and save decoded images to a specified path if the plot directory is not already created.