Add feature extractor training script
In the newly added feature extraction training script, the model is prepared with configurations for training and evaluation. It includes variables, hyperparameters, and functions to load saved models, initialize tensorboard writer, perform pre-training, save models and use callbacks. Furthermore, training data is prepared with samplers and dataloaders, and losses are calculated for each batch of data.