Refactor and enhance baseline training script
The training script for the baseline model is updated - a more accurate name 'EVAL_CONFIGS' is used instead of 'TEST_CONFIGS'. Changes are also made to improve the handling of model checkpoints, including a new 'agent_step_counter' to track progression. Additional code was added to load the latest saved model if exists, otherwise a new model will be created.