Update weight value and refactor training process
The 'dec' weight in the WEIGHTS dictionary has been updated from 0.1 to 1.0 to maintain the uniformity in the weight values. The training process inside the session has been refactored to be encased in a while loop. This ensures that the model training, updates, and saving processes happen for the desired number of epochs. Also, model creation and loading have been moved prior to the start of the epoch counter loop.