Add functionality to plot original and decoded images
The code has been updated to provide an option to plot both original and decoded images in the utils.py file. Previously, we were only saving decoded images, the revised code will now save both decoded and original images. This helps in comparing the quality of decoded images with the originals. The filenames of the saved images are also modified to reflect this change.