Set number of random traps to zero
The previous value for the number of random traps was 5 across all configurations. Now, it has been changed to zero in an effort to simplify the environment and perhaps help the agent to learn more effectively. Given the repetition of this value change, it might also suggest an overall design shift in the simulation environment.