Comment out model checksum calculations
The code modifications mainly include commenting out the function 'model_checksum', which was used to calculate the checksum of the model parameters, along with related calls and print statements in the 'utils.py' file. This alteration implies that these model checksum calculations are no longer considered necessary or are handled differently.