Add sample rate to baseline trainer
Sample rate parameter has been added to the baseline trainer to allow for random sampling of the transition buffer. This parameter is defined as the ratio of the number of samples to the size of the transition buffer. Also, a minor change in progress bar description format has been made.