Update plotting conditions and adjust configuration parameters
The commit revises the conditions under which plotting processes are invoked for decoded images and model representations. It further optimizes the configuration parameters used in baseline training with a feature extractor. The update ensures plots are made when the directory exists, regardless of the existence of the decoder. Furthermore, the new approach also includes the ability to plot model representations for cases where latent dimensions are either 2 or 3.