Reset iterator in utils.py
An issue was encountered where the iterator was not reset before use in the iterable_env function in utils.py. To correct this, a call to iter_reset() has been added before the for loop that uses iterable_env. This ensures that the iterator is properly reset each time the function is called.