Add demo_loss and parameter to compute_loss in feature_model
A demo_loss function was added that calculates the loss between two latent states depending on the policy. A 'demo' weight was added to the weight dictionary in the FeatureNet initialisation. The compute_loss function was also modified to include the demo_loss if a 'demo' loss value is provided.