Disable automatic computation of optimal policy
The automatic computation of the optimal policy has been disabled in the SimpleGridWorld class. It has been replaced with an option to manually trigger the calculation during the reset function. This change gives more control on when the optimal policy calculation should be executed, allowing for more flexibility and potential efficiency improvements in specific usage conditions.