Update data structures and methods for optimal policy calculation
TransitionBuffer and RandomSampler classes have been updated to accommodate the computation of the optimal policies. The changes include an addition of an optional fifth element to the transition tuple to store the comparison result between optimal and next optimal. The 'sample' method in RandomSampler has also been modified to compute the optimal policy when requested.